+++ CHANGELOG.md
@@ -27,6 +27,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fixed a bad type conversion which could cause underflow on a window resize
 - Alacritty now spawns a login shell on macOS, as with Terminal.app and iTerm2
 - Fixed zombie processes sticking around after launching URLs
+- Zero-width characters are now properly rendered without progressing the cursor
 
 ## Version 0.2.3